리스트 코드 정리

+++ src/components/main/List.jsx
@@ -22,8 +22,6 @@ function List() {
   const dispatch = useDispatch();
   const [posts, setPost] = useState([]); //게시물들
   const [page, setPage] = useState(1); //현재 페이지
-  const [isLoading, setLoading] = useState(false); //로깅 스피너
-  const [cnt, setCnt] = useState(1);
   let postsRef = useRef({});
   let loadingRef = useRef(null);
   let pageRef = useRef({});
@@ -80,7 +78,12 @@ function List() {
   }, []);
 
   const onDeleteHandler = (postId) => {
-    dispatch(__deletePost(postId));
+    const result = window.confirm("삭제하시겠습니까?");
+    if (result) {
+      dispatch(__deletePost(postId));
+    } else {
+      return;
+    }
   };
 
   function compareArray(a, b) {
